requirements.txt
pillow==12.1.1
numpy==2.4.3
opencv-python==4.13.0.92
pyyaml==6.0.3
